Implement Stoplight::Wiring::Memory::Backend
